move bulletrecordwrapper. fix a couple test. no new tests
