+++ src/main/java/com/yahoo/bullet/querying/BulletRecordWrapper.java
@@ -3,7 +3,7 @@
  *  Licensed under the terms of the Apache License, Version 2.0.
  *  See the LICENSE file associated with the project for terms.
  */
-package com.yahoo.bullet.querying.evaluators;
+package com.yahoo.bullet.querying;
 
 import com.yahoo.bullet.record.BulletRecord;
 import com.yahoo.bullet.typesystem.TypedObject;
