+++ src/test/java/com/yahoo/bullet/querying/evaluators/FieldEvaluatorTest.java
@@ -5,9 +5,7 @@
  */
 package com.yahoo.bullet.querying.evaluators;
 
-import com.yahoo.bullet.query.expressions.ComplexFieldExpression;
 import com.yahoo.bullet.query.expressions.FieldExpression;
-import com.yahoo.bullet.query.expressions.ValueExpression;
 import com.yahoo.bullet.record.BulletRecord;
 import com.yahoo.bullet.result.RecordBox;
 import com.yahoo.bullet.typesystem.Type;
@@ -56,21 +54,6 @@ public class FieldEvaluatorTest {
         Assert.assertEquals(evaluator.evaluate(record), new TypedObject(Type.INTEGER_MAP, map));
 
         evaluator = new FieldEvaluator(new FieldExpression("aaa", "abc", "def"));
-        Assert.assertEquals(evaluator.evaluate(record), new TypedObject(Type.INTEGER, 5));
-    }
-
-    @Test
-    public void testComplexConstructor() {
-        FieldEvaluator evaluator = new FieldEvaluator(new ComplexFieldExpression("abc", new ValueExpression(0)));
-        Assert.assertEquals(evaluator.evaluate(record), new TypedObject(Type.INTEGER_MAP, map));
-
-        evaluator = new FieldEvaluator(new ComplexFieldExpression("abc", new ValueExpression(0), new ValueExpression("def")));
-        Assert.assertEquals(evaluator.evaluate(record), new TypedObject(Type.INTEGER, 5));
-
-        evaluator = new FieldEvaluator(new ComplexFieldExpression("aaa", new ValueExpression("abc")));
-        Assert.assertEquals(evaluator.evaluate(record), new TypedObject(Type.INTEGER_MAP, map));
-
-        evaluator = new FieldEvaluator(new ComplexFieldExpression("aaa", new ValueExpression("abc"), new ValueExpression("def")));
         Assert.assertEquals(evaluator.evaluate(record), new TypedObject(Type.INTEGER, 5));
     }
 }
